resource_monitoring/python/requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.8.0
exceptiongroup==1.1.1
fastapi==0.98.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pydantic==1.10.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
urllib3==1.26.16
uvicorn==0.22.0
